refactor: Move command interfaces to internal folders. These are not intended for public use
