+++ go/integTest/parallelized_test.go
@@ -7,12 +7,12 @@ import (
 	"time"
 
 	"github.com/google/uuid"
-	"github.com/valkey-io/valkey-glide/go/api"
+	"github.com/valkey-io/valkey-glide/go/internal/interfaces"
 )
 
 func (suite *GlideTestSuite) TestParallelizedSetWithGC() {
 	// The insane 640 parallelism is required to reproduce https://github.com/valkey-io/valkey-glide/issues/3207.
-	suite.runParallelizedWithDefaultClients(640, 640000, 2*time.Minute, func(client api.BaseClientCommands) {
+	suite.runParallelizedWithDefaultClients(640, 640000, 2*time.Minute, func(client interfaces.BaseClientCommands) {
 		runtime.GC()
 		key := uuid.New().String()
 		value := uuid.New().String()
