+++ go/internal/interfaces/scripting_and_function_base_commands.go
@@ -1,6 +1,6 @@
 // Copyright Valkey GLIDE Project Contributors - SPDX Identifier: Apache-2.0
 
-package api
+package interfaces
 
 import "github.com/valkey-io/valkey-glide/go/api/models"
